+++ src/test/java/com/order/api/application/interactors/InsurancePolicyInteractorTest.java
@@ -11,11 +11,9 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import org.mockito.MockitoAnnotations;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 import java.math.BigDecimal;
-import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Map;
 import java.util.UUID;
@@ -41,10 +39,10 @@ class InsurancePolicyInteractorTest {
                 UUID.randomUUID(),
                 UUID.randomUUID(),
                 PolicyCategory.PROPERTY,
-                Map.of("any_coverage", 5000.0),
+                Map.of("any_coverage", BigDecimal.valueOf(5000)),
                 List.of("any_assistance"),
-                50.0,
-                50000.0,
+                BigDecimal.valueOf(50.0),
+                BigDecimal.valueOf(50000.0),
                 PaymentMethod.PIX,
                 SalesChannel.MOBILE
         );
